refactor: 입력 모델 패키지 분리
