+++ src/main/java/me/coonect/coonect/member/adapter/in/web/EmailVerificationController.java
@@ -4,9 +4,9 @@ import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import me.coonect.coonect.member.adapter.in.web.dto.request.EmailVerificationConfirmRequest;
 import me.coonect.coonect.member.adapter.in.web.dto.request.SendVerificationMailRequest;
-import me.coonect.coonect.member.application.port.in.EmailVerificationConfirmCommand;
+import me.coonect.coonect.member.application.port.in.model.EmailVerificationConfirmCommand;
 import me.coonect.coonect.member.application.port.in.EmailVerificationUseCase;
-import me.coonect.coonect.member.application.port.in.SendVerificationEmailCommand;
+import me.coonect.coonect.member.application.port.in.model.SendVerificationEmailCommand;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
