+++ src/main/java/me/coonect/coonect/member/adapter/in/web/dto/request/EmailVerificationConfirmRequest.java
@@ -4,7 +4,7 @@ import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import me.coonect.coonect.member.application.port.in.EmailVerificationConfirmCommand;
+import me.coonect.coonect.member.application.port.in.model.EmailVerificationConfirmCommand;
 
 @Getter
 @AllArgsConstructor
