+++ src/main/java/me/coonect/coonect/member/adapter/in/web/dto/request/SendVerificationMailRequest.java
@@ -4,7 +4,7 @@ import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import me.coonect.coonect.member.application.port.in.SendVerificationEmailCommand;
+import me.coonect.coonect.member.application.port.in.model.SendVerificationEmailCommand;
 
 @Getter
 @AllArgsConstructor
