+++ src/main/java/me/coonect/coonect/member/application/domain/service/EmailVerificationService.java
@@ -6,9 +6,9 @@ import lombok.RequiredArgsConstructor;
 import me.coonect.coonect.common.error.exception.BusinessException;
 import me.coonect.coonect.common.error.exception.ErrorCode;
 import me.coonect.coonect.member.application.domain.exception.EmailDuplicationException;
-import me.coonect.coonect.member.application.port.in.EmailVerificationConfirmCommand;
+import me.coonect.coonect.member.application.port.in.model.EmailVerificationConfirmCommand;
 import me.coonect.coonect.member.application.port.in.EmailVerificationUseCase;
-import me.coonect.coonect.member.application.port.in.SendVerificationEmailCommand;
+import me.coonect.coonect.member.application.port.in.model.SendVerificationEmailCommand;
 import me.coonect.coonect.member.application.port.out.infrastructure.EmailVerificationCodeGenerator;
 import me.coonect.coonect.member.application.port.out.infrastructure.EmailVerificationMailSender;
 import me.coonect.coonect.member.application.port.out.persistence.EmailVerificationCodeRepository;
