+++ src/main/java/me/coonect/coonect/member/application/port/in/EmailVerificationUseCase.java
@@ -1,5 +1,8 @@
 package me.coonect.coonect.member.application.port.in;
 
+import me.coonect.coonect.member.application.port.in.model.EmailVerificationConfirmCommand;
+import me.coonect.coonect.member.application.port.in.model.SendVerificationEmailCommand;
+
 public interface EmailVerificationUseCase {
 
   void sendVerificationEmail(SendVerificationEmailCommand command);
