+++ src/main/java/me/coonect/coonect/member/application/port/in/MemberSignupUseCase.java
@@ -1,6 +1,7 @@
 package me.coonect.coonect.member.application.port.in;
 
 import me.coonect.coonect.member.application.domain.model.Member;
+import me.coonect.coonect.member.application.port.in.model.MemberSignupCommand;
 
 public interface MemberSignupUseCase {
