+++ src/main/java/me/coonect/coonect/member/application/port/in/NicknameValidationUseCase.java
@@ -1,5 +1,7 @@
 package me.coonect.coonect.member.application.port.in;
 
+import me.coonect.coonect.member.application.port.in.model.NicknameValidationQuery;
+
 public interface NicknameValidationUseCase {
 
   boolean isNicknameDuplicated(NicknameValidationQuery query);
