+++ src/main/java/me/coonect/coonect/member/application/port/in/model/EmailVerificationConfirmCommand.java
@@ -1,4 +1,4 @@
-package me.coonect.coonect.member.application.port.in;
+package me.coonect.coonect.member.application.port.in.model;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
@@ -6,14 +6,17 @@ import lombok.Getter;
 import me.coonect.coonect.common.validation.Validator;
 
 @Getter
-public class SendVerificationEmailCommand {
+public class EmailVerificationConfirmCommand {
 
   @Email
   @NotBlank
   String email;
 
-  public SendVerificationEmailCommand(String email) {
+  String code;
+
+  public EmailVerificationConfirmCommand(String email, String code) {
     this.email = email;
+    this.code = code;
 
     Validator.validate(this);
   }
