+++ src/main/java/me/coonect/coonect/member/application/port/in/model/SendVerificationEmailCommand.java
@@ -1,4 +1,4 @@
-package me.coonect.coonect.member.application.port.in;
+package me.coonect.coonect.member.application.port.in.model;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
@@ -6,17 +6,14 @@ import lombok.Getter;
 import me.coonect.coonect.common.validation.Validator;
 
 @Getter
-public class EmailVerificationConfirmCommand {
+public class SendVerificationEmailCommand {
 
   @Email
   @NotBlank
   String email;
 
-  String code;
-
-  public EmailVerificationConfirmCommand(String email, String code) {
+  public SendVerificationEmailCommand(String email) {
     this.email = email;
-    this.code = code;
 
     Validator.validate(this);
   }
