+++ src/main/java/me/coonect/coonect/member/application/port/in/model/validation/Password.java
@@ -1,4 +1,4 @@
-package me.coonect.coonect.member.application.port.in.validation;
+package me.coonect.coonect.member.application.port.in.model.validation;
 
 import static java.lang.annotation.ElementType.FIELD;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
@@ -7,7 +7,7 @@ import jakarta.validation.Constraint;
 import java.lang.annotation.Documented;
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
-import me.coonect.coonect.member.application.port.in.validation.validator.PasswordValidator;
+import me.coonect.coonect.member.application.port.in.model.validation.validator.PasswordValidator;
 
 @Target({FIELD})
 @Retention(RUNTIME)
