+++ src/main/java/me/coonect/coonect/member/application/port/in/model/validation/validator/PasswordValidator.java
@@ -1,8 +1,8 @@
-package me.coonect.coonect.member.application.port.in.validation.validator;
+package me.coonect.coonect.member.application.port.in.model.validation.validator;
 
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
-import me.coonect.coonect.member.application.port.in.validation.Password;
+import me.coonect.coonect.member.application.port.in.model.validation.Password;
 
 public class PasswordValidator implements ConstraintValidator<Password, String> {
