+++ src/test/java/me/coonect/coonect/member/adapter/in/web/dto/request/MemberSignupRequestTest.java
@@ -3,7 +3,7 @@ package me.coonect.coonect.member.adapter.in.web.dto.request;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalDate;
-import me.coonect.coonect.member.application.port.in.MemberSignupCommand;
+import me.coonect.coonect.member.application.port.in.model.MemberSignupCommand;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
