+++ src/test/java/me/coonect/coonect/member/application/domain/service/EmailVerificationServiceTest.java
@@ -8,8 +8,8 @@ import java.time.LocalDate;
 import me.coonect.coonect.common.error.exception.BusinessException;
 import me.coonect.coonect.member.application.domain.exception.EmailDuplicationException;
 import me.coonect.coonect.member.application.domain.model.Member;
-import me.coonect.coonect.member.application.port.in.EmailVerificationConfirmCommand;
-import me.coonect.coonect.member.application.port.in.SendVerificationEmailCommand;
+import me.coonect.coonect.member.application.port.in.model.EmailVerificationConfirmCommand;
+import me.coonect.coonect.member.application.port.in.model.SendVerificationEmailCommand;
 import me.coonect.coonect.member.application.port.out.infrastructure.EmailVerificationCodeGenerator;
 import me.coonect.coonect.member.application.port.out.persistence.EmailVerificationCodeRepository;
 import me.coonect.coonect.member.application.port.out.persistence.MemberRepository;
