+++ src/test/java/me/coonect/coonect/member/application/domain/service/NicknameValidationServiceTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalDate;
 import me.coonect.coonect.member.application.domain.model.Member;
-import me.coonect.coonect.member.application.port.in.NicknameValidationQuery;
+import me.coonect.coonect.member.application.port.in.model.NicknameValidationQuery;
 import me.coonect.coonect.member.application.port.in.NicknameValidationUseCase;
 import me.coonect.coonect.member.application.port.out.persistence.MemberRepository;
 import me.coonect.coonect.mock.FakeMemberRepository;
