+++ src/test/java/me/coonect/coonect/member/application/port/in/model/MemberSignupCommandTest.java
@@ -1,4 +1,4 @@
-package me.coonect.coonect.member.application.port.in;
+package me.coonect.coonect.member.application.port.in.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -7,6 +7,7 @@ import static org.assertj.core.api.Assertions.catchThrowable;
 import jakarta.validation.ConstraintViolationException;
 import java.time.LocalDate;
 import me.coonect.coonect.member.application.domain.model.Member;
+import me.coonect.coonect.member.application.port.in.model.MemberSignupCommand;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
