+++ src/test/java/me/coonect/coonect/member/application/port/in/model/NicknameValidationQueryTest.java
@@ -1,10 +1,11 @@
-package me.coonect.coonect.member.application.port.in;
+package me.coonect.coonect.member.application.port.in.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.Assertions.catchThrowable;
 
 import jakarta.validation.ConstraintViolationException;
+import me.coonect.coonect.member.application.port.in.model.NicknameValidationQuery;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
